chore(daemon,orchestrator): keep receipt owners within the readability ratchet
The S1 fixes pushed command-store.ts and runFactsBuilder.ts past the
600-line cap for non-baseline files. Two natural owners move out whole,
comments intact (no doc squeezing):

- packages/daemon/src/terminal-authority.ts: the pure durable-terminal
  validation shared by the reconciled fast path and artifact recovery
  (recoveredTerminalFacts, terminalCommandResult, ENOENT-tolerant
  lstatOrNull).
- packages/orchestrator/src/runFactsDeliverable.ts: canonicalDeliverable
  with its lstat + real-content probe.

No behavior change; both files are re-exported nowhere and imported by
their single former host.
